renamed auth service and removed redundant code
