+++ src/app/core/services/auth-guard.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, CanActivateChild, Router, RouterStateSnapshot } from '@angular/router';
 
-import { AuthService } from './auth-service';
+import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root',
