+++ src/app/features/change-password/change-password.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 
-import { AuthService } from '@core/services/auth-service';
+import { AuthService } from '@core/services/auth.service';
 import { UserService } from '@core/services/user.service';
 import { Subscription } from 'rxjs';
 import { UserDetails } from '@core/model/userDetails.model';
