+++ src/app/features/change-password/confirmation/confirmation.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthService } from '../../../core/services/auth-service';
+import { AuthService } from '@core/services/auth.service';
 
 @Component({
   selector: 'app-confirmation',
