+++ src/app/features/your-account/your-account.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { UserService } from '../../core/services/user.service';
-import { AuthService } from '@core/services/auth-service';
+import { AuthService } from '@core/services/auth.service';
 import { Subscription } from 'rxjs';
 import { UserDetails } from '@core/model/userDetails.model';
